requirements.txt
asgiref==3.5.0
black==23.9.1
click==8.1.7
Django==4.0.3
djangorestframework==3.14.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
pytz==2023.3.post1
sqlparse==0.4.2
